one place for otelconfig
