+++ otelcollector/prom-config-validator-builder/main.go
@@ -10,47 +10,13 @@ import (
 	"regexp"
 	"strings"
 
+	"github.com/prometheus-collector/shared"
 	"go.opentelemetry.io/collector/confmap"
 	"go.opentelemetry.io/collector/confmap/provider/envprovider"
 	"go.opentelemetry.io/collector/confmap/provider/fileprovider"
 	"go.opentelemetry.io/collector/otelcol"
 	yaml "gopkg.in/yaml.v2"
 )
-
-type OtelConfig struct {
-	Exporters  interface{} `yaml:"exporters"`
-	Processors interface{} `yaml:"processors"`
-	Extensions interface{} `yaml:"extensions"`
-	Receivers  struct {
-		Prometheus struct {
-			Config          interface{} `yaml:"config"`
-			TargetAllocator interface{} `yaml:"target_allocator"`
-			ApiServer       interface{} `yaml:"api_server"`
-		} `yaml:"prometheus"`
-	} `yaml:"receivers"`
-	Service struct {
-		Extensions interface{} `yaml:"extensions"`
-		Pipelines  struct {
-			Metrics struct {
-				Exporters  interface{} `yaml:"exporters"`
-				Processors interface{} `yaml:"processors"`
-				Receivers  interface{} `yaml:"receivers"`
-			} `yaml:"metrics"`
-			MetricsTelemetry struct {
-				Exporters  interface{} `yaml:"exporters,omitempty"`
-				Processors interface{} `yaml:"processors,omitempty"`
-				Receivers  interface{} `yaml:"receivers,omitempty"`
-			} `yaml:"metrics/telemetry,omitempty"`
-		} `yaml:"pipelines"`
-		Telemetry struct {
-			Logs struct {
-				Level       interface{} `yaml:"level"`
-				Encoding    interface{} `yaml:"encoding"`
-				OutputPaths []string    `yaml:"output_paths"`
-			} `yaml:"logs"`
-		} `yaml:"telemetry"`
-	} `yaml:"service"`
-}
 
 var RESET = "\033[0m"
 var RED = "\033[31m"
@@ -93,7 +59,7 @@ func setFatalErrorMessageAsEnvVar(message string) {
 }
 
 func generateOtelConfig(promFilePath string, outputFilePath string, otelConfigTemplatePath string) error {
-	var otelConfig OtelConfig
+	var otelConfig shared.OtelConfig
 
 	otelConfigFileContents, err := ioutil.ReadFile(otelConfigTemplatePath)
 	if err != nil {
