+++ otelcollector/shared/configmap/mp/tomlparser-set-global-config.go
@@ -5,42 +5,9 @@ import (
 	"log"
 	"os"
 
+	"github.com/prometheus-collector/shared"
 	yaml "gopkg.in/yaml.v2"
 )
-
-type OtelConfig struct {
-	Exporters  interface{} `yaml:"exporters"`
-	Processors interface{} `yaml:"processors"`
-	Extensions interface{} `yaml:"extensions"`
-	Receivers  struct {
-		Prometheus struct {
-			Config          map[string]interface{} `yaml:"config"`
-			TargetAllocator interface{}            `yaml:"target_allocator"`
-			APIServer       interface{}            `yaml:"api_server"`
-		} `yaml:"prometheus"`
-	} `yaml:"receivers"`
-	Service struct {
-		Extensions interface{} `yaml:"extensions"`
-		Pipelines  struct {
-			Metrics struct {
-				Exporters  interface{} `yaml:"exporters"`
-				Processors interface{} `yaml:"processors"`
-				Receivers  interface{} `yaml:"receivers"`
-			} `yaml:"metrics"`
-			MetricsTelemetry struct {
-				Exporters  interface{} `yaml:"exporters,omitempty"`
-				Processors interface{} `yaml:"processors,omitempty"`
-				Receivers  interface{} `yaml:"receivers,omitempty"`
-			} `yaml:"metrics/telemetry,omitempty"`
-		} `yaml:"pipelines"`
-		Telemetry struct {
-			Logs struct {
-				Level    interface{} `yaml:"level"`
-				Encoding interface{} `yaml:"encoding"`
-			} `yaml:"logs"`
-		} `yaml:"telemetry"`
-	} `yaml:"service"`
-}
 
 func SetGlobalSettingsInCollectorConfig() {
 	azmonSetGlobalSettings := os.Getenv("AZMON_SET_GLOBAL_SETTINGS")
@@ -53,7 +20,7 @@ func SetGlobalSettingsInCollectorConfig() {
 			return
 		}
 		var promScrapeConfig map[string]interface{}
-		var otelConfig OtelConfig
+		var otelConfig shared.OtelConfig
 		err = yaml.Unmarshal([]byte(mergedCollectorConfigFileContents), &otelConfig)
 		if err != nil {
 			fmt.Printf("Unable to unmarshal merged otel configuration from: %s - %v\n", mergedCollectorConfigFileContents, err)
@@ -71,7 +38,7 @@ func SetGlobalSettingsInCollectorConfig() {
 				fmt.Printf("Unable to read file contents from: %s - %v\n", replicasetCollectorConfigFileContents, err)
 				return
 			}
-			var otelConfigReplicaset OtelConfig
+			var otelConfigReplicaset shared.OtelConfig
 			err = yaml.Unmarshal([]byte(replicasetCollectorConfigFileContents), &otelConfigReplicaset)
 			if err != nil {
 				fmt.Printf("Unable to unmarshal merged otel configuration from: %s - %v\n", replicasetCollectorConfigFileContents, err)
